requirements.txt
openai
requests
python-dotenv==1.0.1
pyTelegramBotAPI
reportlab
aiogram
requests
pydub
openai
pandas
SQLAlchemy
psycopg2-binary